requirements.txt
amqp==5.2.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
Django==5.1.2
django-cors-headers==4.6.0
django-redis==5.4.0
django-rest-authtoken==2.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
msgpack==1.1.0
mysqlclient==2.2.5
permission==0.4.1
pillow==11.0.0
prompt_toolkit==3.0.48
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
service-identity==24.1.0
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.1.0